anolis: block: fix type mismatch build warning
OpenAnolis Bug Tracker: 0000557

Fix the following build warning:
block/blk-zoned.c: In function ‘blk_zone_valid’:
./include/linux/kern_levels.h:5:18: warning: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 3 has type ‘sector_t {aka long unsigned int}’ [-Wformat=]

Fixes: b7bfe33b1bfb ("block: Enhance blk_revalidate_disk_zones()")
Acked-by: Xiaoguang Wang <xiaoguang.wang@linux.alibaba.com>
Acked-by: Xunlei Pang <xlpang@linux.alibaba.com>

+++ block/blk-zoned.c
@@ -470,7 +470,7 @@ static bool blk_zone_valid(struct gendisk *disk, struct blk_zone *zone,
 	/* Check for holes in the zone report */
 	if (zone->start != *sector) {
 		pr_warn("%s: Zone gap at sectors %llu..%llu\n",
-			disk->disk_name, *sector, zone->start);
+			disk->disk_name, (unsigned long long)*sector, zone->start);
 		return false;
 	}
